requirements.txt
appdirs==1.4.4
attrs==22.1.0
autopage==0.5.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cliff==4.0.0
cmd2==2.4.2
cryptography==37.0.4
debtcollector==2.5.0
decorator==5.1.1
dogpile.cache==1.1.8
environs==9.5.0
idna==3.3
importlib-metadata==4.12.0
iso8601==1.0.2
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.3
keystoneauth1==5.0.0
marshmallow==3.17.1
msgpack==1.0.4
munch==2.5.0
netaddr==0.8.0
netifaces==0.11.0
openstacksdk==0.101.0
os-client-config==2.1.0
os-service-types==1.7.0
osc-lib==2.6.2
oslo.config==9.0.0
oslo.context==5.0.0
oslo.i18n==5.1.0
oslo.log==5.0.0
oslo.serialization==5.0.0
oslo.utils==6.0.1
packaging==21.3
pbr==5.10.0
prettytable==3.4.0
pycparser==2.21
pyinotify==0.9.6
pyparsing==3.0.9
pyperclip==1.8.2
python-cinderclient==9.1.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-keystoneclient==5.0.1
python-neutronclient==8.1.0
python-novaclient==18.1.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requestsexceptions==1.4.0
rfc3986==2.0.0
simplejson==3.17.6
six==1.16.0
stevedore==4.0.0
urllib3==1.26.12
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.1
